Refactor (clean up Form.handlers.js, relocate Packing List functions and prepare for outsourcing to component)
